requirements.txt
websocket-client==1.6.3
statsmodels==0.14.1
pandas==2.1.4
numpy==1.24.3
schedule==1.2.0